Use buffer local variable from global hooks
Since I couldn't find a way to set permanent-local-hook property to
generated function, I workaround this by:
(1) defining functions for hooks globally and
(2) using buffer local variables to control the behavior of the
functions.

I don't like the loop in refbuf/reflect-change, which is called via
after-change-functions hook; this may slow down editing. But as I
don't know better solution, use it and see if this implementation
works in practice.